docs: rework filtering docs (#2819)
Update documentation for `useFilters` and `useGlobalFilter` to clarify
what each does, remove old documentation (from v6, seemingly), improve
wording and add `state` details.

Previously, the documentation was hard to follow on why one would use
`useFilters` compared to `useGlobalFilter`.  It was hard to know what
each applied to (be it a row, column(s) or the table as a whole) - in
reality, the former is filtering rows on a per-column basis and the
latter is a single filter value being applied to _all_ columns at once.

These changes augment the existing docs with this info and explain the
other filter hook respectively, aiming to clarify what is being filtered
and by what value.